+++ src/i18n/index.js
@@ -1,4 +1,6 @@
-import { getLocale, isRtl } from '@edx/frontend-platform/i18n';
+import { messages as footerMessages } from '@edx/frontend-component-footer';
+import { messages as headerMessages } from '@edx/frontend-component-header';
+
 import arMessages from './messages/ar.json';
 import deMessages from './messages/de.json';
 import es419Messages from './messages/es_419.json';
@@ -13,7 +15,7 @@ import ukMessages from './messages/uk.json';
 import zhcnMessages from './messages/zh_CN.json';
 // no need to import en messages-- they are in the defaultMessage field
 
-const messages = {
+const appMessages = {
   ar: arMessages,
   'es-419': es419Messages,
   'fa-ir': faIRMessages,
@@ -28,25 +30,8 @@ const messages = {
   uk: ukMessages,
 };
 
-export const getLocalizedSlash = () => {
-  // For fractional grades
-  // if we are in a LTR language, we want to use a forward slash.
-  // If we are in a RTL language, we want to use a backslash instead
-  if (isRtl(getLocale())) {
-    return '\\';
-  }
-  return '/';
-};
-
-export const getLocalizedPercentSign = () => {
-  // LTR languages put the percent to the right of a number.
-  // RTL languages put the percent sign to the left of the number.
-  // We can place a non-printing unicode right-to-left marker next to the percent
-  // sign to make it print to the left of the number if we are currently in a LTR language
-  if (isRtl(getLocale())) {
-    return '\u200f%';
-  }
-  return '%';
-};
-
-export default messages;
+export default [
+  footerMessages,
+  headerMessages,
+  appMessages,
+];
